+++ src/Candoumbe.Pipelines/Components/Docker/IBuildDockerImage.cs
@@ -1,11 +1,9 @@
 ﻿using Nuke.Common;
-using Nuke.Common.IO;
 using Nuke.Common.Tooling;
 using Nuke.Common.Tools.Docker;
 using Nuke.Common.Utilities.Collections;
 
 using System.Collections.Generic;
-using System.Linq;
 
 using static Nuke.Common.Tools.Docker.DockerTasks;
 
@@ -19,12 +17,12 @@ public interface IBuildDockerImage : IHaveConfiguration
     /// <summary>
     /// Gets the docker files to build
     /// </summary>
-    IEnumerable<(AbsolutePath Path, string Name, string Tag)> DockerFiles => Enumerable.Empty<(AbsolutePath, string, string)>();
+    IEnumerable<DockerFile> DockerFiles { get; }
 
     /// <summary>
     /// Build docker images
     /// </summary>
-    public Target Build => _ => _
+    public Target BuildDockerImages => _ => _
         .Description("Build docker images")
         .OnlyWhenStatic(() => DockerFiles.AtLeastOnce())
         .Executes(() =>
@@ -37,14 +35,13 @@ public interface IBuildDockerImage : IHaveConfiguration
                 .Apply(BuildSettings)
                 .CombineWith(DockerFiles,
                              (setting, dockerFile) => setting.SetFile(dockerFile.Path)
-                                                             .WhenNotNull(this as IHaveGitVersion,
-                                                                          (_, version) => _.SetTag($"{dockerFile.Name}:{version.MajorMinorPatchVersion}"))));
+                                                             .SetTag($"{dockerFile.Name}:{dockerFile.Tag}")));
         });
 
     /// <summary>
     /// Default image settings
     /// </summary>
-    public sealed Configure<DockerBuildSettings> BuildSettingsBase => _ => _;
+    public sealed Configure<DockerBuildSettings> BuildSettingsBase => _ => _.SetPath(".");
 
     /// <summary>
     /// Configures the docker image settings