+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Docker.IBuildDockerImage` component
 - Added `IPushDockerImage` component
 - Added `PushDockerImageConfiguration`
+- Added `DockerFile` class
 
 ### ⚠️ Breaking changes
 - Renamed `Candoumbe.Pipelines.IPublish` to `Candoumbe.Pipelines.IPushNugetPackages`